+++ _build/build.config.php
@@ -5,5 +5,5 @@
  */
 define('PKG_NAME', 'modxMCP');
 define('PKG_NAMESPACE', 'modxmcp');
-define('PKG_VERSION', '1.8.11');
+define('PKG_VERSION', '1.8.12');
 define('PKG_RELEASE', 'pl');